Доработка структуры страницы "Каталог продукции"
- блок с товарами и блок с дополнительными товарами вынесены в Partials/catalog/goods и Partials/catalog/additional-goods
