+++ src/pages/_partials/catalog/additional-goods/additional-goods.pug
@@ -1,0 +1,6 @@
+.additional-goods(id='additional-goods')
+  .container
+    h3 Дополнительные товары
+    .additional-goods__inner
+      .temp ojnsvijnsdv
+
