+++ src/pages/_partials/catalog/goods/goods.scss
@@ -1,0 +1,45 @@
+.goods {
+  .container {
+    @media (max-width: $mobile) {
+      padding: 0;
+    }
+  }
+
+  h1 {
+    margin: 46px 0 85px;
+
+    @media (max-width: $mobile) {
+      padding-bottom: 40px;
+      padding-top: 27px;
+      margin-top: 0;
+      margin-bottom: 1px;
+      padding-left: 20px;
+      border-top: 1px solid $grey-button-hover;
+      border-bottom: 1px solid $grey-button-hover;
+    }
+  }
+
+  &__goods {
+    margin-top: 165px;
+    margin-bottom: 75px;
+    display: flex;
+    justify-content: flex-start;
+    align-items: flex-start;
+    flex-wrap: wrap;
+    column-gap: 80px;
+    row-gap: 143px;
+
+    @media (max-width: $tablet) {
+      column-gap: 60px;
+      row-gap: 140px;
+    }
+
+    @media (max-width: $mobile) {
+      row-gap: unset;
+      margin-top: 0;
+    }
+  }
+  @media (max-width: $mobile) {
+    padding-top: 66px;
+  }
+}
